refactor: move tracking_service.dart
